Add files via upload
I has made a new project for TFS2015. I have also edited the code to send a request for TFS2015 to the new project. I has also enclosed the last version of NuGet.exe. Needs to be a part of the project in order to build the latest versions with NuGet.